Refactor: movie list page
