+++ src/pages/Body.jsx
@@ -6,7 +6,7 @@ import Aside from "../components/ui/Aside.jsx";
 import "../features/body/body.css";
 
 
-export default function Body() {
+const Body = () => {
     return (
         <div className="body">
             <Aside></Aside>
@@ -14,3 +14,6 @@ export default function Body() {
         </div>
     );
 };
+
+
+export default Body;
